fix(fitness): workout distance to the centimetre (#89)
distance_m was integer metres, which lost a yard on a 625 yd swim (571.5 m
stored as 572, read back as 626 yd). The column becomes numeric(10,2),
every writer rounds to two places, and readers cast to float8.

+++ modules/fitness/manifest.ts
@@ -35,7 +35,7 @@ export default defineModule({
         kind: z.enum(['strength', 'run', 'ride', 'swim', 'walk', 'other']).default('other'),
         started_at: z.string(),
         duration_s: z.number().int().min(0).max(86_400),
-        distance_m: z.number().int().min(0).max(1_000_000).default(0),
+        distance_m: z.number().min(0).max(1_000_000).default(0),
         avg_hr: z.number().int().min(20).max(260).nullable().optional(),
         sets: z
           .array(
